# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Security-first OpenAPI catalog API for SpecRelay
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
  details?: string[];
}

export interface Workspace {
  id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface WorkspaceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}

export type WorkspaceOverviewRole = typeof WorkspaceOverviewRole[keyof typeof WorkspaceOverviewRole];


export const WorkspaceOverviewRole = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER',
} as const;

export type AuditEventMetadata = { [key: string]: unknown };

export interface AuditEvent {
  id: string;
  workspaceId: string;
  eventType: string;
  /** @nullable */
  resourceType: string | null;
  /** @nullable */
  resourceId: string | null;
  metadata: AuditEventMetadata;
  createdAt: string;
}

export interface WorkspaceOverview {
  workspace: Workspace;
  role: WorkspaceOverviewRole;
  canManage: boolean;
  /** @minimum 0 */
  apiCount: number;
  /** @minimum 0 */
  operationCount: number;
  /** @minimum 0 */
  enabledOperationCount: number;
  recentAuditEvents: AuditEvent[];
}

export type ApiSpecVersionFormat = typeof ApiSpecVersionFormat[keyof typeof ApiSpecVersionFormat];


export const ApiSpecVersionFormat = {
  json: 'json',
  yaml: 'yaml',
} as const;

export type ApiSecuritySchemeType = typeof ApiSecuritySchemeType[keyof typeof ApiSecuritySchemeType];


export const ApiSecuritySchemeType = {
  apiKey: 'apiKey',
  http: 'http',
  unsupported: 'unsupported',
} as const;

/**
 * @nullable
 */
export type ApiSecuritySchemeLocation = typeof ApiSecuritySchemeLocation[keyof typeof ApiSecuritySchemeLocation] | null;


export const ApiSecuritySchemeLocation = {
  header: 'header',
  query: 'query',
} as const;

export interface ApiSecurityScheme {
  name: string;
  type: ApiSecuritySchemeType;
  /** @nullable */
  location: ApiSecuritySchemeLocation;
  /** @nullable */
  parameterName: string | null;
  bearer: boolean;
}

export type SecurityWarningSeverity = typeof SecurityWarningSeverity[keyof typeof SecurityWarningSeverity];


export const SecurityWarningSeverity = {
  info: 'info',
  warning: 'warning',
  blocked: 'blocked',
} as const;

export interface SecurityWarning {
  code: string;
  message: string;
  /** @nullable */
  path: string | null;
  severity: SecurityWarningSeverity;
}

export interface ApiSpecVersion {
  id: string;
  apiId: string;
  version: string;
  format: ApiSpecVersionFormat;
  openapiVersion: string;
  /** @pattern ^[a-f0-9]{64}$ */
  documentHash: string;
  serverUrls: string[];
  securitySchemes: ApiSecurityScheme[];
  validationWarnings: SecurityWarning[];
  importedAt: string;
}

export interface ApiSource {
  id: string;
  workspaceId: string;
  name: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
  latestSpecification: ApiSpecVersion | null;
}

export interface ApiSourceInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @nullable */
  description?: string | null;
}

export type OperationParameterLocation = typeof OperationParameterLocation[keyof typeof OperationParameterLocation];


export const OperationParameterLocation = {
  path: 'path',
  query: 'query',
  header: 'header',
  cookie: 'cookie',
} as const;

export interface OperationParameter {
  name: string;
  location: OperationParameterLocation;
  required: boolean;
  /** @nullable */
  schemaType: string | null;
  /** @nullable */
  description: string | null;
}

export interface RequestBodyMetadata {
  required: boolean;
  contentTypes: string[];
  /** @nullable */
  description: string | null;
}

export interface ResponseMetadata {
  statusCode: string;
  /** @nullable */
  description: string | null;
  contentTypes: string[];
}

export type ApiOperationRisk = typeof ApiOperationRisk[keyof typeof ApiOperationRisk];


export const ApiOperationRisk = {
  READ_LIKE: 'READ_LIKE',
  WRITE: 'WRITE',
  DESTRUCTIVE: 'DESTRUCTIVE',
  UNKNOWN: 'UNKNOWN',
} as const;

export type ApiOperationSecurityGroupsItemItem = {
  scheme: string;
  scopes: string[];
};

export interface ApiOperation {
  id: string;
  apiId: string;
  specificationId: string;
  method: string;
  path: string;
  /** @nullable */
  operationId: string | null;
  displayName: string;
  /** @nullable */
  summary: string | null;
  /** @nullable */
  description: string | null;
  tags: string[];
  parameters: OperationParameter[];
  requestBody: RequestBodyMetadata | null;
  responses: ResponseMetadata[];
  securityRequirements: string[];
  securityGroups: ApiOperationSecurityGroupsItemItem[][];
  risk: ApiOperationRisk;
  enabled: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ApiDetail {
  api: ApiSource;
  operations: ApiOperation[];
  canManage: boolean;
}

export interface SpecificationImportInput {
  /** @maxLength 255 */
  filename?: string;
  /**
     * @minLength 2
     * @maxLength 2097152
     */
  document: string;
}

export interface ValidationSummary {
  valid: boolean;
  warnings: SecurityWarning[];
  unsupported: string[];
}

export interface ImportResult {
  specification: ApiSpecVersion;
  operations: ApiOperation[];
  validation: ValidationSummary;
}

export interface OperationStateUpdate {
  enabled: boolean;
}

export type CredentialMetadataType = typeof CredentialMetadataType[keyof typeof CredentialMetadataType];


export const CredentialMetadataType = {
  API_KEY: 'API_KEY',
  BEARER: 'BEARER',
} as const;

export type CredentialMetadataLocation = typeof CredentialMetadataLocation[keyof typeof CredentialMetadataLocation];


export const CredentialMetadataLocation = {
  header: 'header',
  query: 'query',
} as const;

export type CredentialMetadataStatus = typeof CredentialMetadataStatus[keyof typeof CredentialMetadataStatus];


export const CredentialMetadataStatus = {
  ACTIVE: 'ACTIVE',
  DISABLED: 'DISABLED',
  REVOKED: 'REVOKED',
} as const;

export interface CredentialMetadata {
  id: string;
  schemeName: string;
  type: CredentialMetadataType;
  location: CredentialMetadataLocation;
  parameterName: string;
  label: string;
  configured: boolean;
  status: CredentialMetadataStatus;
  destinationHost: string;
  createdAt: string;
  updatedAt: string;
}

export interface CredentialInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  schemeName: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 8192
     */
  secret: string;
}

export type McpRequestMethod = typeof McpRequestMethod[keyof typeof McpRequestMethod];


export const McpRequestMethod = {
  'server/discover': 'server/discover',
  'tools/list': 'tools/list',
  'tools/call': 'tools/call',
} as const;

export type McpRequestParams = { [key: string]: unknown };

export interface McpRequest {
  jsonrpc: '2.0';
  id: string | number | null;
  method: McpRequestMethod;
  params: McpRequestParams;
}

/**
 * JSON-RPC response containing either a tool list or tool call result.
 */
export interface McpResponse { [key: string]: unknown }

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;