# lib/api-zod/src/generated/types/apiOperation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Security-first OpenAPI catalog API for SpecRelay
 * OpenAPI spec version: 0.1.0
 */
import type { ApiOperationRisk } from './apiOperationRisk';
import type { ApiOperationSecurityGroupsItemItem } from './apiOperationSecurityGroupsItemItem';
import type { OperationParameter } from './operationParameter';
import type { RequestBodyMetadata } from './requestBodyMetadata';
import type { ResponseMetadata } from './responseMetadata';

export interface ApiOperation {
  id: string;
  apiId: string;
  specificationId: string;
  method: string;
  path: string;
  /** @nullable */
  operationId: string | null;
  displayName: string;
  /** @nullable */
  summary: string | null;
  /** @nullable */
  description: string | null;
  tags: string[];
  parameters: OperationParameter[];
  requestBody: RequestBodyMetadata | null;
  responses: ResponseMetadata[];
  securityRequirements: string[];
  securityGroups: ApiOperationSecurityGroupsItemItem[][];
  risk: ApiOperationRisk;
  enabled: boolean;
  createdAt: Date;
  updatedAt: Date;
}